Switch to using elapsed time to throttle rendering
Hopefully more precise than trying to estimate the amount of drawing we can do, and won’t need to be tweaked every time the rendering algorithm is improved.

There are separate, hard-coded limits for the simulator and for the actual device, based on experience. Currently using separate .jungle files to make that work.